requirements.txt
google-generativeai
torch
transformers
numpy
pandas
scikit-learn
tqdm
python-dotenv
datasets
huggingface_hub
Pillow
matplotlib
seaborn